Add ability to process custom icon symbols from BaseCamp. Fix QuickAction bug.

+++ OsmAnd/src/net/osmand/plus/quickaction/QuickAction.java
@@ -128,7 +128,7 @@ public class QuickAction {
     }
 
 	public long getType() {
-		return type;
+		return type.toInt();
 	}
 
     public boolean isActionEditable() {
